#24 refactor: App Stack 리팩토링
